Restore cloned pattern sheets (#63)
After all connectors were bound to InkScape with a buggy implementation. With coloured dots instructions are even easier for clones. More over: the tile next to each patch looks a bit like the [legend] for the advanced options of the web page. The dots replace the digits, identical coloured dots should move together.

[legend]: https://github.com/d-bl/GroundForge/blob/46e04ee/images/legend.png