Twist Markers in Inkscape are too long (#193)
Markers in Inkscape are too long.

There is a bug in Inkscape where it is not using the "viewbox" attribute to clip the size of the marker:  https://bugs.launchpad.net/inkscape/+bug/1690952/comments/7
Workaround is to change the SVG twist marker to have length 4 instead of length 12.  The current length of 12 is unnecessarily long.